requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.3
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-compat==1.0.15
django-extensions==3.2.3
django-redis==5.3.0
djangorestframework==3.14.0
executing==1.2.0
Faker==18.10.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kombu==5.3.2
matplotlib-inline==0.1.6
mysqlclient==2.1.0
nest-asyncio==1.5.6
numpy==1.26.1
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
redis==5.0.0
requests==2.31.0
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wheel==0.41.2
zope.interface==6.0
